extract some statements
