+++ lib/handlers/propertyAccessorHandlers.js
@@ -1,5 +1,7 @@
 var tokenUtils = require("../util/tokenUtils");
 
+var propertyAccessors = ["giv", "levl"];
+
 /**
  * Handles property accession with an index:
  *  <object> levl <key>
@@ -54,7 +56,15 @@ function handleGiv(parseContext) {
   return ".";
 }
 
+/**
+ * Determines whether the token is a property accessor token
+ */
+function isPropertyAccessor(token) {
+  return propertyAccessors.includes(token);
+}
+
 module.exports = {
   handleLevl,
-  handleGiv
+  handleGiv,
+  isPropertyAccessor
 };
